+++ src/main/java/org/thema/graphab/links/SpacePathFinder.java
@@ -45,6 +45,15 @@ public interface SpacePathFinder {
     public List<double[]> calcPaths(Coordinate p, List<Coordinate> dests);
     
     /**
+     * Calcule les distances à partir du point p vers toutes les 
+     * destinations dests en restant à l'intérieur de la tache
+     * @param p the origin coordinate
+     * @param dests the destination coordinates
+     * @return cost and length of paths between p and ech dests
+     */
+    public List<double[]> calcPathsInsidePatch(Coordinate p, List<Coordinate> dests);
+    
+    /**
      * Calcule les chemins à partir du point p vers tous les patch dont la distance cout est inférieure ou égale à maxCost
      * @param p the origin coordinate
      * @param maxCost maximal distance, zero for no maximum
@@ -63,22 +72,6 @@ public interface SpacePathFinder {
      */
     public HashMap<DefaultFeature, Path> calcPaths(Geometry geom, double maxCost, boolean realPath);
 
-    /**
-     * Calculates the paths from oPatch to all other patches
-     * if all == false, calculates for patches where id is greater than oPatch id
-     * @param oPatch the origin patch
-     * @param realPath keep the real path or just a straight line between centroid patches ?
-     * @return for each destination patch the path from oPatch
-     */
-    public HashMap<Feature, Path> calcPaths(Feature oPatch, double maxCost, boolean realPath, boolean all);
-
-    /**
-     * Calculates the paths from oPatch to all dPatch
-     * @param oPatch the origin patch
-     * @param dPatch the destinations patches
-     * @return for each destination patch the path from oPatch
-     */
-    public HashMap<Feature, Path> calcPaths(Feature oPatch, Collection<Feature> dPatch);
     
     /**
      * Calc nearest patch from point p